refactor: Move SafeEthClient to safeclient package
