+++ aggregator/rollup_writer.go
@@ -15,8 +15,8 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 
 	registryrollup "github.com/NethermindEth/near-sffl/contracts/bindings/SFFLRegistryRollup"
-	"github.com/NethermindEth/near-sffl/core"
 	"github.com/NethermindEth/near-sffl/core/config"
+	"github.com/NethermindEth/near-sffl/core/safeclient"
 	"github.com/NethermindEth/near-sffl/core/types/messages"
 )
 
@@ -45,7 +45,7 @@ func NewRollupWriter(
 	address common.Address,
 	logger logging.Logger,
 ) (*RollupWriter, error) {
-	client, err := core.NewSafeEthClient(rollupInfo.RpcUrl, logger)
+	client, err := safeclient.NewSafeEthClient(rollupInfo.RpcUrl, logger)
 	if err != nil {
 		return nil, err
 	}
