+++ core/metricable.go
@@ -5,6 +5,8 @@ import (
 	sdklogging "github.com/Layr-Labs/eigensdk-go/logging"
 	rpccalls "github.com/Layr-Labs/eigensdk-go/metrics/collectors/rpc_calls"
 	"github.com/prometheus/client_golang/prometheus"
+
+	"github.com/NethermindEth/near-sffl/core/safeclient"
 )
 
 type Metricable interface {
@@ -15,8 +17,8 @@ func CreateEthClientWithCollector(id, url string, enableMetrics bool, registry *
 	if enableMetrics {
 		// Using url as avsName
 		rpcCallsCollector := rpccalls.NewCollector(id+url, registry)
-		return NewSafeEthClient(url, logger, WithCollector(rpcCallsCollector))
+		return safeclient.NewSafeEthClient(url, logger, safeclient.WithCollector(rpcCallsCollector))
 	}
 
-	return NewSafeEthClient(url, logger)
+	return safeclient.NewSafeEthClient(url, logger)
 }
