+++ core/safeclient/client.go
@@ -1,4 +1,4 @@
-package core
+package safeclient
 
 import (
 	"context"
